Phase out post expression (#843)
* Change expression definition

* Rename getParameters to getParameterDeclarations to distinguish the method in FunctionExpression

* Add tests

* Rename parameters to args for clear meaning

* Add javadoc

* Apply indent

* Support window function

* Move ORDER-BY/LIMIT to outermost

* Fix query expression builder

* Avoid error at FE side

* Add timestamp column to ts query

* Fix ts for window function

* Fix legacy IAggregatableColumn support

* Move extends aggregate functions out of Function class

* Fix GROUP-BY

* Fix {interval} macro

* Fix ClassCastException

* Remove Function from query ast

* Remove old implementation

* Fix output name

* Add indent to column list

* Fix col name

* Move Expression2SqlSerializer into builder

* Fix support of ck

* Improve test cases to use fixed timezone

* Fix CI

* Rename toSelectColumn to toSelector

* Disable optimization for aggregation function

* Add test for count aggregate function

* Fix count aggregator

* Remove reference of AggregateCountColumn

* validateParameters ---> validateArgs

* Fix ORDER-BY validation

* Rename GeneralQueryRequest/QueryResponse to QueryRequest/QueryResponse

* Remove FieldExpressionVisitorAdaptor2

+++ server/web-service/src/main/java/org/bithon/server/web/service/datasource/api/TimeSeriesQueryResult.java
@@ -59,7 +59,7 @@ public class TimeSeriesQueryResult {
 
     public static TimeSeriesQueryResult build(TimeSpan start,
                                               TimeSpan end,
-                                              int intervalSecond,
+                                              long intervalSecond,
                                               List<Map<String, Object>> dataPoints,
                                               String tsColumn,
                                               List<String> groups,
@@ -74,7 +74,7 @@ public class TimeSeriesQueryResult {
 
         long startSecond = start.toSeconds() / intervalSecond * intervalSecond;
         long endSecond = end.toSeconds() / intervalSecond * intervalSecond;
-        int bucketCount = (int) (endSecond - startSecond) / intervalSecond + (isEndInclusive ? 1 : 0);
+        int bucketCount = (int) ((endSecond - startSecond) / intervalSecond + (isEndInclusive ? 1 : 0));
 
         // Use LinkedHashMap to retain the order of input metric list
         Map<List<String>, TimeSeriesMetric> map = new LinkedHashMap<>(7);
@@ -89,7 +89,7 @@ public class TimeSeriesQueryResult {
         } else {
             for (Map<String, Object> point : dataPoints) {
                 long timestamp = ((Number) point.get(tsColumn)).longValue();
-                int bucketIndex = (int) (timestamp - startSecond) / intervalSecond;
+                int bucketIndex = (int) ((timestamp - startSecond) / intervalSecond);
 
                 for (String metric : metrics) {
                     // this code is not so efficient
